+++ src/main/java/com/wanted/preonboarding/ticket/application/common/exception/PaymentFailedException.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.ticket.application.exception;
+package com.wanted.preonboarding.ticket.application.common.exception;
 
 public class PaymentFailedException extends ApiException {
     public PaymentFailedException(ExceptionStatus exceptionStatus) {
