+++ src/main/java/com/wanted/preonboarding/ticket/application/common/exception/SeatNotAvailableException.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.ticket.application.exception;
+package com.wanted.preonboarding.ticket.application.common.exception;
 
 public class SeatNotAvailableException extends ApiException {
     public SeatNotAvailableException(ExceptionStatus exceptionStatus) {
