+++ src/main/java/com/wanted/preonboarding/ticket/domain/dto/response/PerformanceDetail.java
@@ -1,6 +1,6 @@
 package com.wanted.preonboarding.ticket.domain.dto.response;
 
-import com.wanted.preonboarding.ticket.application.exception.EntityNotFoundException;
+import com.wanted.preonboarding.ticket.application.common.exception.EntityNotFoundException;
 import com.wanted.preonboarding.ticket.domain.entity.Performance;
 import com.wanted.preonboarding.ticket.domain.entity.PerformanceSeatInfo;
 import lombok.Builder;
@@ -8,7 +8,7 @@ import lombok.Value;
 
 import java.util.List;
 
-import static com.wanted.preonboarding.ticket.application.exception.ExceptionStatus.NOT_FOUND_INFO;
+import static com.wanted.preonboarding.ticket.application.common.exception.ExceptionStatus.NOT_FOUND_INFO;
 import static com.wanted.preonboarding.ticket.application.common.util.TimeFormatter.convertToReadableFormat;
 import static com.wanted.preonboarding.ticket.domain.dto.response.PerformanceInfo.convertToName;
 import static com.wanted.preonboarding.ticket.domain.enums.ReservationAvailability.AVAILABLE;
